структура приложенияя
