+++ Navigation/InfoViewController.swift
@@ -6,7 +6,7 @@
 //
 
 import UIKit
-
+import StorageService
 class InfoViewController: UIViewController {
 
     override func viewDidLoad() {
@@ -42,14 +42,4 @@ class InfoViewController: UIViewController {
         }
     }
 
-    /*
-    // MARK: - Navigation
-
-    // In a storyboard-based application, you will often want to do a little preparation before navigation
-    override func prepare(for segue: UIStoryboardSegue, sender: Any?) {
-        // Get the new view controller using segue.destination.
-        // Pass the selected object to the new view controller.
-    }
-    */
-
-
+
