--- Navigation/Post.swift
@@ -1,11 +1,0 @@
-//
-//  Post.swift
-//  Navigation
-//
-//  Created by Никита  on 05.06.2023.
-//
-
-import Foundation
-struct Post {
-    var title:String
-}
