+++ Navigation/ViewController.swift
@@ -11,7 +11,7 @@ class ViewController: UIViewController {
 
    
 
-       override func viewDidLoad() {
+      public override func viewDidLoad() {
            super.viewDidLoad()
            title = "Profile"
