+++ StorageService/StorageService.h
@@ -1,0 +1,18 @@
+//
+//  StorageService.h
+//  StorageService
+//
+//  Created by Никита  on 24.07.2023.
+//
+
+#import <Foundation/Foundation.h>
+
+//! Project version number for StorageService.
+FOUNDATION_EXPORT double StorageServiceVersionNumber;
+
+//! Project version string for StorageService.
+FOUNDATION_EXPORT const unsigned char StorageServiceVersionString[];
+
+// In this header, you should import all the public headers of your framework using statements like #import <StorageService/PublicHeader.h>
+
+
